+++ R/RcppExports.R
@@ -61,7 +61,7 @@ microtex_add_font_from_otf <- function(otf_path, index = 0L) {
     .Call(`_gridmicrotex_microtex_add_font_from_otf`, otf_path, index)
 }
 
-parse_latex_cpp <- function(tex, text_size = 20.0, line_space = 10.0, fg_color = "#000000", max_width = 0, math_font = "", main_font = "", use_path = TRUE, tex_style = "", justify = FALSE) {
-    .Call(`_gridmicrotex_parse_latex_cpp`, tex, text_size, line_space, fg_color, max_width, math_font, main_font, use_path, tex_style, justify)
+parse_latex_cpp <- function(tex, text_size = 20.0, line_space = 10.0, fg_color = "#000000", max_width = 0, math_font = "", main_font = "", use_path = TRUE, tex_style = "", justify = FALSE, optimal_break = FALSE) {
+    .Call(`_gridmicrotex_parse_latex_cpp`, tex, text_size, line_space, fg_color, max_width, math_font, main_font, use_path, tex_style, justify, optimal_break)
 }